Interpolate using piecewise polynomials (pp) with coeffs computed on the fly in vertsolve
This avoids excessive memory requirements of pre-computing PPCs for the entire ocean, and is faster than interpolating on the fly (with the `interp1d` subpackage) where pp coeffs are computed each time you want to evaluate the interpolant.  Since vertsolve will do tons of evaluations of the interpolant (many within the same interval), avoiding this recomputation of pp coeffs greatly improves execution time.