v2.2.0 -- BFS wetting for initial surface. Stabilize S, T.
- omega_surf no longer has `ref` or `pin_p` arguments. To initialize it
from a potential density or specific volume anomaly surface, pre-calc
that and pass that as `p_init`. If `p_init` is given as a float, instead
of an array of floats, the initial surface is now constructed by BFS
wetting.

- Stability module. Provides functions to check for statically unstable
casts, and for making minimal perturbations to ensure static stability.